exp changed to long & max level pet

+++ src/main/kotlin/dev/dungeoncrawler/data/PetLevelData.kt
@@ -1,3 +1,3 @@
 package dev.dungeoncrawler.data
 
-class PetLevelData(val levels: HashMap<String, Pair<Int, Int>> = HashMap())
+class PetLevelData(val levels: HashMap<String, Pair<Int, Long>> = HashMap())
